requirements.txt
cffi==1.9.1
cryptography==1.7.1
falcon==1.0.0
google-api-python-client==1.5.5
html2text==2016.9.19
httplib2==0.9.2
hug==2.2.0
idna==2.1
Jinja2==2.8
keyring==10.1
mailchimp3==2.0.3
Markdown==2.6.7
MarkupSafe==0.23
marrow.mailer==4.0.2
marrow.util==1.2.3
oauth2client==4.0.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.17
python-mimeparse==1.6.0
requests==2.12.4
rsa==3.4.2
SecretStorage==2.3.1
six==1.10.0
uritemplate==3.0.0
voluptuous==0.9.3
whitenoise==3.2.2
